requirements.txt
# Use Python 3.11, 3.12, or 3.13 for this project.
# Hugging Face datasets currently breaks on Python 3.14 during dataset loading.
# Core dependency for downloading and preprocessing the receipt datasets.
datasets[vision]>=2.19.0,<4.0.0
transformers[torch]>=4.41.0,<5.0.0
accelerate>=0.30.0
torch>=2.2.0
Pillow>=10.0.0
sentencepiece>=0.2.0
protobuf>=4.25.0
# LayoutLMv3 needs numpy + the LayoutLMv3 image processor stack.
numpy>=1.24
# pytesseract drives the Tesseract binary for predict_layoutlmv3.py.
# Install Tesseract separately: macOS `brew install tesseract`, Ubuntu `apt-get install tesseract-ocr`.
pytesseract>=0.3.10
